scripts/model_pipeline/requirements.txt
amadeus==7.0.0
boto3
google-cloud-bigquery
google-cloud-storage
db-types
pandas
python-dotenv
requests